% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggdiagram}
\alias{ob_point}
\alias{ob_ellipse}
\alias{ob_rectangle}
\alias{ob_label}
\alias{ob_circle}
\alias{place}
\alias{connect}
\alias{set_default_arrowhead}
\alias{arrow_head_line}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggarrow}{\code{\link[ggarrow:arrow_ornaments]{arrow_head_line}}}

  \item{ggdiagram}{\code{\link[ggdiagram]{connect}}, \code{\link[ggdiagram]{ggdiagram}}, \code{\link[ggdiagram]{ob_circle}}, \code{\link[ggdiagram]{ob_ellipse}}, \code{\link[ggdiagram]{ob_label}}, \code{\link[ggdiagram]{ob_point}}, \code{\link[ggdiagram]{ob_rectangle}}, \code{\link[ggdiagram]{place}}, \code{\link[ggdiagram:arrowhead]{set_default_arrowhead}}}
}}

