% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_dag}
\alias{theme_dag}
\title{A clean theme for DAG diagrams}
\usage{
theme_dag()
}
\value{
A \link[ggplot2:theme]{ggplot2::theme} object.
}
\description{
Based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_void()}} with centered bold title and subtitle.
}
\examples{
library(daggr)
theme_dag()
}
