% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daggr-package.R
\docType{package}
\name{daggr-package}
\alias{daggr}
\alias{daggr-package}
\title{daggr: Quick DAG Diagrams with Sensible Defaults}
\description{
Wraps 'ggdiagram' and 'ggarrow' with opinionated defaults for drawing directed acyclic graphs (DAGs). Provides styled node constructors, a connection helper, and a clean theme so that 'library(daggr)' replaces ~30 lines of boilerplate setup.
}
\author{
\strong{Maintainer}: Benjamin Guinaudeau \email{benjamin.guinaudeau@gmail.com}

}
\keyword{internal}
